Refactor VM build to improve control and reliability
This adds the ability to specify multiple disk volumes, CPU cores, and
memory settings for a VM.  It also reduces some of the redundancy in
ansible variables defining a VM (through better use of machine groups).
In addition, it improves the ansible configuration process for VMs
by only initializing the VMs after the host machines have been
initialized, blocking ansible from progressing until the OS install
has finished (avoiding having it error when trying to perform the next
steps on a VM that isn't ready yet), and will perform the OS install in
parallel on multiple VMs (significantly speeding up builds when multiple
VMs need to be rebuilt).

Unfortunately there are still some intermittent failures that appear to
be race conditions associated with the kubernetes setup steps.