Accept false negative change message for manifests
When applying kubernetes manifests, an output stating 'configured'
sometimes indicates a change, but also can be seen every time certain
manifests are applied. Originally we were reporting these as changes,
but that meant that every ansible run reported changes and it became
easy to ignore the change message.  Now we do not flag 'configured' as a
change, so all ansible runs without changes will report no changes (but
a few rare corner cases with changes will not report changes).